replaced flutter gauge package with syncfusion
swapped out the old FlutterGauge widget with an SFReadialGauge.
added a 30 day trial license key to prevent the popup.

need to get the meter looking and behaving the way we want, once thats done we need to start feeding bools from a list constantly from the time the user hits the start recorder button.